requirements.txt
alembic==1.13.1
aniso8601==9.0.1
attrs==23.2.0
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.2
Flask-Migrate==4.0.5
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
Mako==1.3.2
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2024.1
referencing==0.33.0
rpds-py==0.17.1
SQLAlchemy==2.0.25
typing_extensions==4.9.0
Werkzeug==3.0.1
